feat: add send timeout if server is not available
If the elastic server ist not available every request blocks execution.
Now if the server is not reachable the events get saved in a queue and sent with the next request but with a defined timeout. As long the timeout has not exeeded the target do not try to send the request to the server.